carbon/requirements.txt
celery
SQLAlchemy
django
djangorestframework
mysql-connector-python
mysqlclient
bcrypt
djangorestframework-simplejwt
#python-dotenv